# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenAPI natation project backend documentation",
        "description": "backend is used by smartphone application and desktop front",
        "contact": {
            "email": "**"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/post-exercice": {
            "post": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Submit an exercise",
                "description": "Endpoint to submit an exercise",
                "operationId": "postExercice",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "programme_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "exercise_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Exercice 1"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "gps": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "latitude": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 48.856614
                                                },
                                                "longitude": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 2.352222
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful exercise submission",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid input"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/performance/record": {
            "post": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Record exercise performance from mobile",
                "description": "Enregistrer les performances depuis mobile (annexe 9-1-B)",
                "operationId": "recordPerformance",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "athlete_id": {
                                        "description": "ID de l'athlète",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "exercise_id": {
                                        "description": "ID de l'exercice",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "program_id": {
                                        "description": "ID du programme",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "duration": {
                                        "description": "Durée réalisée en secondes",
                                        "type": "integer",
                                        "example": 1800
                                    },
                                    "distance": {
                                        "description": "Distance parcourue",
                                        "type": "number",
                                        "format": "float",
                                        "example": 2000.5
                                    },
                                    "repetitions": {
                                        "description": "Répétitions effectuées",
                                        "type": "integer",
                                        "example": 25
                                    },
                                    "weight": {
                                        "description": "Poids utilisé",
                                        "type": "number",
                                        "format": "float",
                                        "example": 75.5
                                    },
                                    "heart_rate": {
                                        "description": "Fréquence cardiaque moyenne",
                                        "type": "integer",
                                        "example": 155
                                    },
                                    "calories": {
                                        "description": "Calories brûlées",
                                        "type": "integer",
                                        "example": 320
                                    },
                                    "notes": {
                                        "description": "Notes de l'athlète",
                                        "type": "string",
                                        "example": "Exercice hors piscine réalisé"
                                    },
                                    "gps_data": {
                                        "description": "Données GPS pour exercices hors piscine",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "latitude": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 48.856614
                                                },
                                                "longitude": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 2.352222
                                                },
                                                "timestamp": {
                                                    "type": "string",
                                                    "example": "2024-01-01T12:00:00Z"
                                                },
                                                "speed": {
                                                    "description": "Vitesse en km/h",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 12.5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Performance recorded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Performance enregistrée avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid input"
                                        },
                                        "details": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
    },
...
}